Extract internal/retrieval shared core for vault search
