+++ internal/retrieval/entity.go
@@ -1,4 +1,4 @@
-package store
+package retrieval
 
 import (
 	"regexp"
