+++ internal/retrieval/synonyms_test.go
@@ -1,4 +1,4 @@
-package store
+package retrieval
 
 import (
 	"strings"
