+++ internal/store/store_test.go
@@ -599,22 +599,3 @@ func TestMmapPragmaSet(t *testing.T) {
 	require.NoError(t, err)
 	assert.Greater(t, mmapSize, int64(0), "mmap_size should be non-zero after init")
 }
-
-// --- Stopwords ---
-
-func TestIsStopword(t *testing.T) {
-	// Common English
-	assert.True(t, IsStopword("the"))
-	assert.True(t, IsStopword("update"))
-	// FTS5/search-domain terms
-	assert.True(t, IsStopword("porter"))
-	assert.True(t, IsStopword("trigram"))
-	assert.True(t, IsStopword("prose"))
-	// Common data field names
-	assert.True(t, IsStopword("name"))
-	assert.True(t, IsStopword("title"))
-	assert.True(t, IsStopword("id"))
-	// Real terms should not be stopwords
-	assert.False(t, IsStopword("authentication"))
-	assert.False(t, IsStopword("middleware"))
-}
